Script to copy Tendenci installs
Being able to make a quick clone of a site for testing can be handy and if
resources are tight using the same server might be useful.
Other reasons to do this might be some form of a/b deployment or updating a dev
site from test.

Requires some configuration to be performed.